Use is_unsupported_term on both unix and windows
